requirements.txt
altgraph==0.16.1
astroid==2.3.3
asyncio==3.4.3
autopep8==1.4.4
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nest-asyncio==1.2.3
numpy==1.18.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyInstaller==3.6
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyserial==3.4
pyserial-asyncio==0.4
pytz==2019.3
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
